Moved picker pages related to sharing to a 'pickers' subfolder
